third_party/blink/web_tests/platform/generic/fast/dom/StyleSheet/gc-declaration-parent-rule-expected.txt

Test that CSS rules don't lose custom properties on GC when only reachable through its declarations.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS style.parentRule.foo is "bar"
PASS successfullyParsed is true

TEST COMPLETE
